feat: allow customize role for registration and disable/enable registration
