--- app/Panel/Conference/Livewire/AccessSetting.php
@@ -1,61 +1,0 @@
-<?php
-
-namespace App\Panel\Conference\Livewire;
-
-use App\Facades\Setting;
-use Filament\Forms\Components\Actions;
-use Filament\Forms\Components\Actions\Action;
-use Filament\Forms\Components\Checkbox;
-use Filament\Forms\Components\Section;
-use Filament\Forms\Concerns\InteractsWithForms;
-use Filament\Forms\Contracts\HasForms;
-use Filament\Forms\Form;
-use Livewire\Component;
-
-class AccessSetting extends Component implements HasForms
-{
-    use InteractsWithForms;
-
-    public ?array $formData = [];
-
-    public function mount()
-    {
-        $this->form->fill(Setting::all());
-    }
-
-    public function render()
-    {
-        return view('forms.form');
-    }
-
-    public function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Section::make()
-                    ->schema([
-                        Checkbox::make('allow_registration')
-                            ->label('Allow Registration')
-                            ->helperText('Allow public to register on the site.'),
-                    ])
-                    ->columns(1),
-                Actions::make([
-                    Action::make('save')
-                        ->label('Save')
-                        ->successNotificationTitle('Saved!')
-                        ->failureNotificationTitle('Data could not be saved.')
-                        ->action(function (Action $action) {
-                            $formData = $this->form->getState();
-                            try {
-                                Setting::update($formData);
-
-                                $action->sendSuccessNotification();
-                            } catch (\Throwable $th) {
-                                $action->failure();
-                            }
-                        }),
-                ])->alignLeft(),
-            ])
-            ->statePath('formData');
-    }
-}
